Move magic link async request to a struct `MagicLinkRequester` for reuse in the password screen.
